extract duplicated string encryption into a shared helper class ShellSecureInput
